pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-ethics-multiagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Marvins20", email = "marcusabrantes1337@gmail.com" }
]
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "langchain>=0.3.18",
    "langchain-core>=0.3.31",
    "langchain-community>=0.3.18",
    "langchain-google-genai>=2.0.8",
    "langchain-chroma>=0.1.0",
    "langchain-text-splitters>=0.3.0",
    "chromadb>=0.6.5",
    "python-dotenv>=1.0.0",
    "rank-bm25>=0.2.2",
    "gql[httpx]>=4.0.0",
    "httpx>=0.28.1",
    "langgraph>=1.0.7",
    "mcp[cli]>=1.26.0",
    "deepagents>=0.4.1",
    "unstructured[doc,docs,docx,pdf,txt]>=0.20.8",
    "duckdb>=1.4.4",
    "pandas>=3.0.1",
]

[project.scripts]
ai-ethics-multiagents = "src.main:running_agent"

[tool.hatch.build.targets.wheel]
packages = ["src"]
